+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 ### Fix
+
 - different query start times with multiple queries due to mutating of grafana start time
 
 ## 1.1.1