+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 ### Added
+
 - Default Percentile Value when percentile metric is selected
 
 ## 1.1.1